+++ src/main/java/ai/mcpdirect/gateway/dao/entity/aitool/AIPortTool.java
@@ -1,0 +1,17 @@
+package ai.mcpdirect.gateway.dao.entity.aitool;
+
+
+public class AIPortTool {
+    public long id;
+    public long makerId;
+    public int status;
+    public long lastUpdated;
+    public String name;
+    public int hash;
+    public String metaData;
+    public String tags;
+    public long agentId;
+    public int agentStatus;
+    public int makerStatus;
+    public AIPortTool() {}
+}