+++ src/main/java/ai/mcpdirect/gateway/dao/mapper/account/AccessKeyMapper.java
@@ -1,0 +1,24 @@
+package ai.mcpdirect.gateway.dao.mapper.account;
+
+import ai.mcpdirect.gateway.dao.entity.account.AIPortAccessKeyCredential;
+import org.apache.ibatis.annotations.*;
+
+import java.util.List;
+
+public interface AccessKeyMapper {
+    String accessKeyTable="account.access_key";
+    String selectAccessKeyCredential = """
+            SELECT id,
+            secret_key secretKey,
+            status,"name",
+            effective_date effectiveDate,
+            expiration_date expirationDate,
+            user_id userId,
+            user_roles userRoles,
+            usage_amount usageAmount,
+            created FROM
+            """+accessKeyTable+"\n";
+
+    @Select(selectAccessKeyCredential+"where id=#{accessKeyId}")
+    AIPortAccessKeyCredential selectAccessKeyCredentialById(@Param("accessKeyId") long accessKeyId);
+}