+++ src/main/java/ai/mcpdirect/gateway/dao/mapper/account/AccountMapper.java
@@ -1,0 +1,9 @@
+package ai.mcpdirect.gateway.dao.mapper.account;
+
+import org.apache.ibatis.annotations.*;
+
+@Mapper
+public interface AccountMapper extends AccessKeyMapper,TeamMapper{
+    String userTable = "account.user";
+    String userAccountTable = "account.user_account";
+}