+++ src/main/java/ai/mcpdirect/gateway/util/AIToolDirectory.java
@@ -7,8 +7,10 @@ import java.util.*;
 public class AIToolDirectory {
     public static class Description{
         public String name;
-        public List<String> tags;
+//        public List<String> tags;
+        public String tags;
         public ServiceDescription metaData;
+//        public MCPToolSchema metaData;
     }
     public static class Tools{
         public String engineId;
@@ -16,16 +18,11 @@ public class AIToolDirectory {
     }
 
     public long userId;
-//    public String user;
-    public String secretKey;
-//    public HashSet<Tool> tools = new HashSet<>();
     public Map<String,Tools> tools = new HashMap<>();
 
-    public static AIToolDirectory create(long id, String secretKey){
+    public static AIToolDirectory create(long id){
         AIToolDirectory provider = new AIToolDirectory();
         provider.userId=id;
-//        provider.user = user;
-        provider.secretKey = secretKey;
         return provider;
     }
 }